WIP: Weapon, Tank, Viewport classes

+++ PaperTanksV2-Client/PageStates/PageState.cs
@@ -7,7 +7,7 @@ namespace PaperTanksV2Client.PageStates
     {
         void init(Game game);
         void input(Game game);
-        void update(Game game, double deltaTime);
+        void update(Game game, float deltaTime);
         void prerender(Game game, SKCanvas canvas, RenderStates renderStates);
         void render(Game game, SKCanvas canvas, RenderStates renderStates);
         void postrender(Game game, SKCanvas canvas, RenderStates renderStates);
